more fixes
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/branches/pcl-1.x@4877 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ sample_consensus/include/pcl/sample_consensus/impl/prosac.hpp
@@ -168,7 +168,7 @@ pcl::ProgressiveSampleConsensus<PointT>::computeModel (int debug_verbosity_level
           break;
 
         // If we find a better epsilon_n_star
-        float epsilon_possible_n_star = (float)I_possible_n_star / possible_n_star;
+        float epsilon_possible_n_star = static_cast<float>(I_possible_n_star) / static_cast<float>(possible_n_star);
         // Make sure we have a better epsilon_possible_n_star
         if ((epsilon_possible_n_star > epsilon_n_star) && (epsilon_possible_n_star > epsilon_possible_n_star_best))
         {
